backend/requirements.txt
asgiref==3.11.0
Django==6.0.1
django-cleanup==9.0.0
django-colorfield==0.14.0
django-cors-headers==4.9.0
django-rest-knox==5.0.2
django-unfold==0.75.0
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
pillow==12.1.0
pip==25.3
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.11.0
